requirements.txt
streamlit
torch
torchvision
huggingface_hub
pillow
numpy

